# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  models.Specification:
    enum:
    - Frontend
    - Backend
    - Machine Learning
    - Developer
    - Designer
    - Manager
    type: string
    x-enum-varnames:
    - Frontend
    - Backend
    - MachineLearning
    - Developer
    - Designer
    - Manager
  models.User:
    properties:
      about:
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      passwordHash:
        type: string
      photoURL:
        type: string
      resumeURL:
        type: string
      specification:
        $ref: '#/definitions/models.Specification'
      telegram:
        type: string
      updatedAt:
        type: string
    type: object
  models.UserRole:
    properties:
      id:
        type: string
      roleID:
        type: string
      userID:
        type: string
    type: object
host: 109.73.202.151:8080
info:
  contact: {}
  description: ITaM API
  title: LiveCode API
  version: "1.0"
paths:
...
  /auth/api/get_user/{user_id}:
    get:
      description: Возвращает данные текущего пользователя
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid user ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить информацию о пользователе
      tags:
      - User
...
  /auth/api/get_user_properties:
    get:
      description: Возвращает список свойств текущего пользователя
      produces:
      - application/json
      responses:
        "200":
          description: User properties
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Получить свойства пользователя
      tags:
      - User
  /auth/api/get_user_roles:
    get:
      description: Возвращает список ролей текущего пользователя
      produces:
      - application/json
      responses:
        "200":
          description: User roles
          schema:
            items:
              $ref: '#/definitions/models.UserRole'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Получить роли пользователя
      tags:
      - User
  /auth/api/login:
    post:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      description: Авторизация пользователя с использованием логина и пароля
      parameters:
      - description: Login credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Логин пользователя
      tags:
      - User
  /auth/api/me:
    get:
      description: Возвращает данные авторизованного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Получить информацию о текущем пользователе
      tags:
      - User
...
  /auth/api/register:
    post:
      consumes:
      - application/json
      description: Регистрация нового пользователя в системе
      parameters:
      - description: User registration details
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success message with user data
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация нового пользователя
      tags:
      - User
...
  /auth/api/update_user_info:
    patch:
      consumes:
      - application/json
      description: Обновляет профиль пользователя
      parameters:
      - description: User update data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Обновить информацию пользователя
      tags:
      - User
schemes:
- http
securityDefinitions:
  OAuth2Password:
    flow: password
    tokenUrl: /api/login
    type: oauth2
swagger: "2.0"